New solution

+++ 0x01-variables_if_else_while/10-print_comb2.c
@@ -6,14 +6,14 @@
  */
 int main(void)
 {
-	int a, b;
+	float a, b;
 
 	for (a = '0'; a <= '9'; ++a)
 	{
 		for (b = '1'; b <= '9'; ++b)
 		{
-				putchar(a);
-				putchar(b);
+			putchar(a);
+			putchar(b);
 			if (a != '9' || b != '9')
 			{
 				putchar(' ');
